ajax-editing-model-state/KendoGridAjaxEditing/Infrastructure/Search/Implementation/IdFilter.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Web;
using KendoGridAjaxEditing.Infrastructure.Search.Enums;
using KendoGridAjaxEditing.Infrastructure.Search.Interfaces;
using KendoGridAjaxEditing.Models;

namespace KendoGridAjaxEditing.Infrastructure.Search.Implementation
{
    public class IdFilter : ISearchFilter<BinViewModel, SearchBinFilters>
    {
        public IQueryable<BinViewModel> Apply(IQueryable<BinViewModel> query, SearchBinFilters filter)
        {
            if (query != null && filter != null && filter.Id != null && filter.Id.Values != null)
            {
                query = query.Where(b => filter.Id.Values.Any(v => b.ID.Equals(v)));
            }

            return query;
        }
    }
}